// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  DOCTOR: 'DOCTOR',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Species = {
  DOG: 'DOG',
  CAT: 'CAT',
  CATTLE: 'CATTLE',
  BUFFALO: 'BUFFALO',
  SHEEP: 'SHEEP',
  GOAT: 'GOAT',
  HORSE: 'HORSE',
  OTHER: 'OTHER'
} as const

export type Species = (typeof Species)[keyof typeof Species]


export const BlogStatus = {
  Pending: 'Pending',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type BlogStatus = (typeof BlogStatus)[keyof typeof BlogStatus]


export const AppointmentStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]